% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methode_de_Monte_Carlo_en_R.R
\name{monte_carlo_integration}
\alias{monte_carlo_integration}
\title{monte carlo integration}
\usage{
monte_carlo_integration(f, a, b, n)
}
\arguments{
\item{f}{fonction a integrer}

\item{a}{borne inferieur}

\item{b}{borne superieur}

\item{n}{nombre pour la precision}
}
\value{
integration
}
\description{
monte carlo integration
}
