% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methode_des_rectangles_1_en_R.R
\name{rectangle_integration}
\alias{rectangle_integration}
\title{rectangle_1_integration}
\usage{
rectangle_integration(f, a, b, n)
}
\arguments{
\item{f}{fonction a integrer}

\item{a}{borne inferieur}

\item{b}{borne superieur}

\item{n}{nombre pour la precision}
}
\value{
integration
}
\description{
rectangle_1_integration
}
