% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methode_des_rectangles_2_en_R.R
\name{rectangles_integration}
\alias{rectangles_integration}
\title{rectangles_integration}
\usage{
rectangles_integration(f, a, b, n)
}
\arguments{
\item{f}{fonction a integrer}

\item{a}{borne inferieur}

\item{b}{borne superieur}

\item{n}{nombre pour la precision}
}
\value{
integration
}
\description{
rectangles_integration
}
