% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methode_de_Simpson_en_R.R
\name{simpson_integration}
\alias{simpson_integration}
\title{simpson integration}
\usage{
simpson_integration(f, a, b, n)
}
\arguments{
\item{f}{fonction a integrer}

\item{a}{borne inferieur}

\item{b}{borne superieur}

\item{n}{nombre pour la precision}
}
\value{
integration
}
\description{
simpson integration
}
