% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methode_du_trapeze_en_R.R
\name{trapeze_integration}
\alias{trapeze_integration}
\title{trapeze integration}
\usage{
trapeze_integration(f, a, b, n)
}
\arguments{
\item{f}{fonction a integrer}

\item{a}{borne inferieur}

\item{b}{borne superieur}

\item{n}{nombre pour la precision}
}
\value{
integration
}
\description{
Méthode du trapèze pour l'intégration numérique
}
